refactoring of the repo with respect to cpu code separation
clippy warnings fixed
