+++ benches/cpu_sloth.rs
